Initial Fix for modifying Loop Mirroring
Turning off the mirror flag on a loop would cause a crash. This fix adds necessary variables to check whether this flag has been changed, and if so it will copy the mirrored loop into the current loop so it can be displayed correctly.

This needs to be further modified for switching between mirror flags fluently. Also the modification of the x-yHot variables needed to be disabled for some reason for this to work. These should be fixed as well.